refactor: home layout refactoring
