--- app/_features/home/footer/index.ts
@@ -1,1 +1,0 @@
-export { default } from '@/_features/home/footer/footer';
