--- app/_features/home/header/index.ts
@@ -1,1 +1,0 @@
-export { default } from '@/_features/home/header/header';
